float64 chenged to float32

+++ calculator/calculator.go
@@ -1,6 +1,6 @@
 package calculator
 
 // float finalMark = (float) (currentAverageIntFMLayout * 0.6 + examIntFMLayout * 0.4);
-func CalculateFinal(averageMark int, examMark int) float64 {
-	return float64(float64(averageMark)*0.6 + float64(examMark)*0.4)
+func CalculateFinal(averageMark int, examMark int) float32 {
+	return float32(float32(averageMark)*0.6 + float32(examMark)*0.4)
 }
